0.9.6

+++ src/modules/player/inactivityDisconnect.js
@@ -25,24 +25,27 @@ module.exports = class InactivityDisconnect extends Modules {
 
         const guild = this.client.guilds.cache.get(guildId);
 
+        this.client.database.db("guilds").collection("players").deleteOne({ guildId: guildId });
+        this.client.database.db("queues").collection(guildId).deleteMany({});
+
+        try { existingConnection.state.subscription.player.stop(); } catch (e) { }
+
+        clearTimeout(existingConnection.playTimer); clearTimeout(existingConnection.pauseTimer); clearTimeout(existingConnection.aloneTimer);
+        existingConnection.destroy();
+
         if (playerData?.channelId) {
 
-            const announcesChannel = this.client.channels.cache.get(playerData.channelId);
-
-            try { let lastNowPlayingMessage = await announcesChannel.messages.fetch(playerData.messageId); await lastNowPlayingMessage.delete() } catch (e) { };
+            const announcesChannel = await this.client.channels.fetch(playerData.announcesId).catch(e => { });
+            if (!announcesChannel || !announcesChannel.permissionsFor(this.client.user.id).has("SEND_MESSAGES") || !announcesChannel.permissionsFor(this.client.user.id).has("EMBED_LINKS")) { return; };
 
             const inactivityEmbed = new MessageEmbed({ color: guild.me.displayHexColor })
                 .setDescription(`I left the voice channel because I was inactive for too long.\nIf you are a **Premium** member, you can disable this by typing \`/247\`.`)
 
             announcesChannel.send({ embeds: [inactivityEmbed] }).then(m => { setTimeout(() => { m.delete().catch(e => { }) }, 120000) })
 
+            this.client.player.updateNpMessage(guildId, "delete");
+
         }
-
-        this.client.database.db("guilds").collection("players").deleteOne({ guildId: guildId });
-        this.client.database.db("queues").collection(guildId).deleteMany({});
-
-        try { existingConnection.state.subscription.player.stop(); } catch (e) { }
-        existingConnection.destroy();
 
     }
